refactor: move bench metric-policy-preset types to homeboy-extension-contract
BenchMetricPolicyPreset + BenchMetricPolicyPresetKind (and the pure to_policy
impl, which only maps self fields into a BenchMetricPolicy) move to the contract
crate. Their expansion functions (using gate/parsing/error) stay in core.

This clears the last non-self field dependency of BenchResults - all four
top-level bench result types (BenchResults, BenchRunMetadata, BenchScenario,
BenchRunSnapshot) now reference only contract/leaf types.
